requirements.txt
asgiref==3.8.1
confluent-kafka==2.4.0
Django==5.0.6
djangorestframework==3.15.2
kafka-python==2.0.2
pika==1.3.2
sqlparse==0.5.0
